ref(crons): Split tasks
