+++ src/sentry/conf/server.py
@@ -734,7 +734,10 @@ CELERY_IMPORTS = (
     "sentry.incidents.tasks",
     "sentry.snuba.tasks",
     "sentry.replays.tasks",
-    "sentry.monitors.tasks",
+    "sentry.monitors.tasks.clock_pulse",
+    "sentry.monitors.tasks.check_missed",
+    "sentry.monitors.tasks.check_timeout",
+    "sentry.monitors.tasks.detect_broken_monitor_envs",
     "sentry.tasks.ai_autofix",
     "sentry.tasks.app_store_connect",
     "sentry.tasks.assemble",
